Extract Windows1253 to its own file
