+++ escpos4k/src/commonMain/kotlin/cz/multiplatform/escpos4k/core/encoding/charset/Windows1253.kt
@@ -1,0 +1,152 @@
+/*
+ *    Copyright 2024 Ondřej Karmazín
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package cz.multiplatform.escpos4k.core.encoding.charset
+
+/** Page 47 [Windows-1253] */
+public data object Windows1253 : Charset(47, "windows-1253") {
+  override val mapping: Map<Char, Int> =
+      mapOf(
+          '\u20AC' to 0x80,
+          '\uFFFD' to 0x81,
+          '\u201A' to 0x82,
+          '\u0192' to 0x83,
+          '\u201E' to 0x84,
+          '\u2026' to 0x85,
+          '\u2020' to 0x86,
+          '\u2021' to 0x87,
+          '\uFFFD' to 0x88,
+          '\u2030' to 0x89,
+          '\uFFFD' to 0x8A,
+          '\u2039' to 0x8B,
+          '\uFFFD' to 0x8C,
+          '\uFFFD' to 0x8D,
+          '\uFFFD' to 0x8E,
+          '\uFFFD' to 0x8F,
+          '\uFFFD' to 0x90,
+          '\u2018' to 0x91,
+          '\u2019' to 0x92,
+          '\u201C' to 0x93,
+          '\u201D' to 0x94,
+          '\u2022' to 0x95,
+          '\u2013' to 0x96,
+          '\u2014' to 0x97,
+          '\uFFFD' to 0x98,
+          '\u2122' to 0x99,
+          '\uFFFD' to 0x9A,
+          '\u203A' to 0x9B,
+          '\uFFFD' to 0x9C,
+          '\uFFFD' to 0x9D,
+          '\uFFFD' to 0x9E,
+          '\uFFFD' to 0x9F,
+          '\u00A0' to 0xA0,
+          '\u0385' to 0xA1,
+          '\u0386' to 0xA2,
+          '\u00A3' to 0xA3,
+          '\u00A4' to 0xA4,
+          '\u00A5' to 0xA5,
+          '\u00A6' to 0xA6,
+          '\u00A7' to 0xA7,
+          '\u00A8' to 0xA8,
+          '\u00A9' to 0xA9,
+          '\uFFFD' to 0xAA,
+          '\u00AB' to 0xAB,
+          '\u00AC' to 0xAC,
+          '\u00AD' to 0xAD,
+          '\u00AE' to 0xAE,
+          '\u2015' to 0xAF,
+          '\u00B0' to 0xB0,
+          '\u00B1' to 0xB1,
+          '\u00B2' to 0xB2,
+          '\u00B3' to 0xB3,
+          '\u0384' to 0xB4,
+          '\u00B5' to 0xB5,
+          '\u00B6' to 0xB6,
+          '\u00B7' to 0xB7,
+          '\u0388' to 0xB8,
+          '\u0389' to 0xB9,
+          '\u038A' to 0xBA,
+          '\u00BB' to 0xBB,
+          '\u038C' to 0xBC,
+          '\u00BD' to 0xBD,
+          '\u038E' to 0xBE,
+          '\u038F' to 0xBF,
+          '\u0390' to 0xC0,
+          '\u0391' to 0xC1,
+          '\u0392' to 0xC2,
+          '\u0393' to 0xC3,
+          '\u0394' to 0xC4,
+          '\u0395' to 0xC5,
+          '\u0396' to 0xC6,
+          '\u0397' to 0xC7,
+          '\u0398' to 0xC8,
+          '\u0399' to 0xC9,
+          '\u039A' to 0xCA,
+          '\u039B' to 0xCB,
+          '\u039C' to 0xCC,
+          '\u039D' to 0xCD,
+          '\u039E' to 0xCE,
+          '\u039F' to 0xCF,
+          '\u03A0' to 0xD0,
+          '\u03A1' to 0xD1,
+          '\uFFFD' to 0xD2,
+          '\u03A3' to 0xD3,
+          '\u03A4' to 0xD4,
+          '\u03A5' to 0xD5,
+          '\u03A6' to 0xD6,
+          '\u03A7' to 0xD7,
+          '\u03A8' to 0xD8,
+          '\u03A9' to 0xD9,
+          '\u03AA' to 0xDA,
+          '\u03AB' to 0xDB,
+          '\u03AC' to 0xDC,
+          '\u03AD' to 0xDD,
+          '\u03AE' to 0xDE,
+          '\u03AF' to 0xDF,
+          '\u03B0' to 0xE0,
+          '\u03B1' to 0xE1,
+          '\u03B2' to 0xE2,
+          '\u03B3' to 0xE3,
+          '\u03B4' to 0xE4,
+          '\u03B5' to 0xE5,
+          '\u03B6' to 0xE6,
+          '\u03B7' to 0xE7,
+          '\u03B8' to 0xE8,
+          '\u03B9' to 0xE9,
+          '\u03BA' to 0xEA,
+          '\u03BB' to 0xEB,
+          '\u03BC' to 0xEC,
+          '\u03BD' to 0xED,
+          '\u03BE' to 0xEE,
+          '\u03BF' to 0xEF,
+          '\u03C0' to 0xF0,
+          '\u03C1' to 0xF1,
+          '\u03C2' to 0xF2,
+          '\u03C3' to 0xF3,
+          '\u03C4' to 0xF4,
+          '\u03C5' to 0xF5,
+          '\u03C6' to 0xF6,
+          '\u03C7' to 0xF7,
+          '\u03C8' to 0xF8,
+          '\u03C9' to 0xF9,
+          '\u03CA' to 0xFA,
+          '\u03CB' to 0xFB,
+          '\u03CC' to 0xFC,
+          '\u03CD' to 0xFD,
+          '\u03CE' to 0xFE,
+          '\uFFFD' to 0xFF,
+      )
+}
